Finished dubugging on Maxwell.
It turns out not to work well on the case I have, probably because of the nature of the noise in single-shot images.
Details are in ~/Ar2p/testLcurve.ipynb . Max curvature landed in the over-regularized regime, regardless of lsmooth.
But I discovered an issue with the absolute scale of AtA that is previously stored in precontracted.npz , which is fixed.