Implemented a util function for contraction: adaptive_contraction. Unit test passed, as shown in unittest/precontract.ipynb
It is more robust and systematic than its predecessor utils.dict_innerprod . It also chunks up the data to contract when memory is ran out.
I am considering to replace all usage of utils.dict_innerprod with adaptive_contraction in the next iteration of spook. Even so, this adaptive_contraction is easy to use and robust outside the spook instantiator.