Fixed a bug in SpookLinSolve. Tested working with unittest/test1Dsps.py
When the QP problem cannot be separated into sub-problems, the dimensions of the unknown X are flattened.
In setting up the solver like that, in method SpookLinSolve.__setupProbFlat, the P matrix that governs the quadratic term was implemented as a block-sparse-row matrix
which does not support += modification.
Thus I converted the P matrix into CSC format.